Add files via upload
I found that the Intersection function does not even notice a PLANE so we can't see if the math is right. I commented out some of the old equations there and started looking at RAYCAST for why it does not notice the PLANE. Will try and work on this tomorrow.